Move the integration test to the relevant class
