+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/servicebus/domains/envelopes/model/Classification.java
@@ -4,5 +4,5 @@ public enum Classification {
     EXCEPTION,
     NEW_APPLICATION,
     SUPPLEMENTARY_EVIDENCE,
-    SUPPLEMENTARY_EVIDENCE_WITH_OCR
+    SUPPLEMENTARY_EVIDENCE_WITH_OCR,
 }
